+++ FrontEnd/src/App.js
@@ -4,7 +4,7 @@ import './global.css';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import AdminPage from './components/adminPage/AdminPage';
 import BasicPage from './components/basicPage/BasicPage';
-import { AuthProvider } from './hooks';
+import { AuthProvider } from './context';
 
 function App() {
